nvme-print: Move JSON printing function into a new file
Let's split the JSON printing part from the console printing code as
the file is getting to big.
